move split to ts41
